Moving even more into facade
