+++ src/main/java/com/telenordigital/prime/firebase/FbStorage.java
@@ -3,7 +3,8 @@ package com.telenordigital.prime.firebase;
 import com.google.firebase.FirebaseApp;
 import com.google.firebase.FirebaseOptions;
 import com.google.firebase.auth.FirebaseCredentials;
-import com.google.firebase.database.*;
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.FirebaseDatabase;
 import com.telenordigital.prime.events.*;
 import com.telenordigital.prime.ocs.state.OcsState;
 import org.slf4j.Logger;
@@ -16,7 +17,6 @@ import java.util.Collection;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.CountDownLatch;
-import java.util.function.Consumer;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
@@ -49,19 +49,10 @@ public final class FbStorage implements Storage {
         // Load subscriber balance from firebase to in-memory OcsState
         loadSubscriberBalanceDataFromFirebaseToInMemoryStructure(ocsState);
 
-
-
         // Scoop up products left and right (and don't worry about duplicates, race conditions or
         // anything else by sending them to the listeners.
         facade.addProductCatalogListener(this::addOrUpdateProduct);
-
-
-        final ValueEventListener productCatalogValueEventListener = newCatalogDataChangedEventListener(
-                item ->  addTopupProduct(item.getSku(), item.getNoOfBytes())
-        );
-
-
-        facade.addProductCatalogValueListener(productCatalogValueEventListener);
+        facade.addProductCatalogValueListener(item ->  addTopupProduct(item.getSku(), item.getNoOfBytes()));
         facade.addPurchaseEventListener(newChildListenerThatDispatchesPurchaseRequestToExecutor());
     }
 
@@ -101,25 +92,14 @@ public final class FbStorage implements Storage {
     }
 
 
-    private ValueEventListener newCatalogDataChangedEventListener(final Consumer<ProductCatalogItem> consumer) {
-        checkNotNull(consumer);
-        return new ValueEventListener() {
-            @Override
-            public void onDataChange(DataSnapshot snapshot) {
-                LOG.info("onDataChange");
-                interpretDataSnapshotAsProductCatalogItem(snapshot,consumer);
-            }
 
-            @Override
-            public void onCancelled(DatabaseError error) {
-            }
-        };
-    }
+
 
     private long getMillisSinceEpoch() {
         return Instant.now().toEpochMilli();
     }
 
+    // XXX Too  specific!
     private void addOrUpdateProduct(final DataSnapshot snapshot) {
         if (snapshotIsInvalid(snapshot)) return;
 
@@ -192,21 +172,7 @@ public final class FbStorage implements Storage {
         }
     }
 
-    private void interpretDataSnapshotAsProductCatalogItem(final DataSnapshot snapshot, Consumer<ProductCatalogItem> consumer) {
-        checkNotNull(consumer);
-        if (snapshotIsInvalid(snapshot)) return;
 
-        try {
-            final ProductCatalogItem item =
-                    snapshot.getValue(ProductCatalogItem.class);
-            if (item.getSku() != null) {
-               consumer.accept(item);
-            }
-            LOG.info("Just read a product catalog item: " + item);
-        } catch (Exception e) {
-            LOG.error("Couldn't transform req into FbPurchaseRequest", e);
-        }
-    }
 
 
     @Override
